Make log path configurable
Allow log path to be configured so it can be set to a directory backed by a mounted volume in docker. It doesn't seem to be possible to mount a volume under %TEMP%. This way, we can use a service on the host to ship logs, instead of having to integrate it into the container.