Add RHINO_ prefix to environment variables
Match prefix of new RHINO_TOKEN and RHINO_COMPUTE_KEY variables.

* COMPUTE_BIND_URLS → RHINO_COMPUTE_URLS
* COMPUTE_LOG_PATH → RHINO_COMPUTE_LOG_PATH
* COMPUTE_LOG_RETAIN_DAYS → RHINO_COMPUTE_LOG_RETAIN_DAYS
* COMPUTE_BACKEND_PORT ✘ (ignored)

Old variables will still be loaded with a warning, except
COMPUTE_BACKEND_PORT which is now ignored (with a warning).

All of the configuration has been refactored into a single class to keep the
rest of the code clear of deprecated variable handling.