Listen on localhost only by default
Previously would listen on 0.0.0.0 if compiled as Release, which was a
pain if you're running compute locally on your Windows 10 machine
without the correct permissions configured.

This change means that the RHINO_COMPUTE_URLS env var needs to be set to
something like "http://+:80" in order for compute to be able to receive
requests from other computers. It's also necessary to have the correct
permissions and open the ports in the Windows firewall, both of which
are handled by the bootstrap script.

Closes #109