Fire document added event when setting up definition
Add definition to `DocumentServer` to raise `DocumentAdded` event. This event is used by some plug-ins as a way to register handlers for `GH_Document` events when the doc is loaded.

Primarily this is being added to support plug-ins which interact with `SolutionStart` and `SolutionEnd` to do multi-stage solves.

This overload of `AddDocument` is very basic and I haven't seen anything weird so far...!