Update computegen
Migrates project to dotnet core.

Outputs client code and docs to language dirs in clients/. The plan is
for the language dirs in clients/ to replace the computeclient_py and
computeclient_js repos. In addition to the generated code and docs,
they'll also contain non-generated content and configuration files for
packaging and, hopefully, tests too!